test(cli): split dispatch basics suite
